api/requirements.txt
fastapi>=0.115.0,<1.0.0
sqlmodel>=0.0.19,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
redis>=5.0.0,<6.0.0
httpx>=0.25.0,<1.0.0
